Refactor: Decompose combat page into CombatSetupView and ActiveCombatView components, and useCombat hook. Fix: Make auth logout idempotent to pass integration tests.
