+++ app/combat/page.tsx
@@ -1,2410 +1,18 @@
 'use client';
-
-import { useState, useEffect, useRef, useMemo } from 'react';
-import Link from 'next/link';
 import { ProtectedRoute } from '@/lib/components/ProtectedRoute';
-import { CreatureStatBlock } from '@/lib/components/CreatureStatBlock';
-import { QuickCombatantModal } from '@/lib/components/QuickCombatantModal';
-import { CombatInfoIcon } from '@/lib/components/CombatInfoIcon';
+import { ActiveCombatView } from '@/lib/components/ActiveCombatView';
+import { CombatSetupView } from '@/lib/components/CombatSetupView';
 import { useAuth } from '@/lib/hooks/useAuth';
-import { CombatState, CombatantState, Encounter, Character, Party, StatusCondition, InitiativeRoll, Monster, MonsterTemplate, ActiveDamageEffect } from '@/lib/types';
-import { applyDamage as calcApplyDamage, applyHealing as calcApplyHealing, setTempHp as calcSetTempHp, resetIncomingLegendaryPool, sortCombatants, buildLairCombatant, buildCombatantFromSource, applyDamageWithType as calcApplyDamageWithType, mergeActiveDamageEffects, removeActiveDamageEffects, getDexInitiativeBonus, buildInitiativeRoll } from '@/lib/utils/combat';
-import { DAMAGE_TYPE_GROUPS, DAMAGE_EFFECT_PRESETS, DamageType } from '@/lib/constants';
-import { LairForm } from '@/lib/components/LairForm';
-import { LegendaryActionsPanel } from '@/lib/components/LegendaryActionsPanel';
-import { LairActionsSlot } from '@/lib/components/LairActionsSlot';
-import { resolveCharactersForCombat } from '@/lib/utils/partySelection';
-import { processRoundEnd } from '@/lib/combat/conditionExpiry';
-import { pushHpHistory, popHpHistory, getHpHistoryStack, clearCombatHistory } from '@/lib/utils/hpHistory';
-
+import { useCombat } from '@/lib/hooks/useCombat';
 function CombatContent() {
+  const combat = useCombat();
   const { user } = useAuth();
-  const [combatState, setCombatState] = useState<CombatState | null>(null);
-  const [encounters, setEncounters] = useState<Encounter[]>([]);
-  const [characters, setCharacters] = useState<Character[]>([]);
-  const [monsterTemplates, setMonsterTemplates] = useState<MonsterTemplate[]>([]);
-  const [selectedEncounterId, setSelectedEncounterId] = useState<string>('');
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [initiativeMode, setInitiativeMode] = useState(false);
-  const [initiativeFilter, setInitiativeFilter] = useState<'all' | 'player' | 'monster'>('all');
-  const [showQuickEntryType, setShowQuickEntryType] = useState<'player' | 'monster' | null>(null);
-  const [showCombatantModal, setShowCombatantModal] = useState(false);
-  const [setupCombatants, setSetupCombatants] = useState<CombatantState[]>([]);
-  const [parties, setParties] = useState<Party[]>([]);
-  const [selectedPartyId, setSelectedPartyId] = useState<string | null>(null);
-  const [loadingTemplates, setLoadingTemplates] = useState(false);
-  const [selectedDetailCombatantId, setSelectedDetailCombatantId] = useState<string | null>(null);
-  const [detailPosition, setDetailPosition] = useState<{top: number, left: number} | null>(null);
-  const [initiativeEditId, setInitiativeEditId] = useState<string | null>(null);
-  const [removeConfirmId, setRemoveConfirmId] = useState<string | null>(null);
-  const [removeConfirmPosition, setRemoveConfirmPosition] = useState<{top: number, left: number} | null>(null);
-  const [showEncounterDescription, setShowEncounterDescription] = useState(false);
-  const [toast, setToast] = useState<{message: string, type: 'success' | 'error'} | null>(null);
-  const [showLairForm, setShowLairForm] = useState(false);
-  const [lairFormName, setLairFormName] = useState('');
-  const [lairFormSeedMonster, setLairFormSeedMonster] = useState('');
-  const initiativePanelRef = useRef<HTMLDivElement>(null);
-  const setupCombatantsRef = useRef<CombatantState[]>([]);
-
-  useEffect(() => {
-    const loadData = async () => {
-      try {
-        setLoading(true);
-        setLoadingTemplates(true);
-        setError(null);
-        const [encountersRes, charactersRes, combatRes, monstersRes, partiesRes] = await Promise.all([
-          fetch('/api/encounters'),
-          fetch('/api/characters'),
-          fetch('/api/combat'),
-          fetch('/api/monsters'),
-          fetch('/api/parties'),
-        ]);
-
-        if (!encountersRes.ok || !charactersRes.ok || !combatRes.ok || !monstersRes.ok || !partiesRes.ok) {
-          throw new Error('Failed to load data');
-        }
-
-        const encountersData = await encountersRes.json();
-        const charactersData = await charactersRes.json();
-        const combatData = await combatRes.json();
-        const monstersData = await monstersRes.json();
-        const partiesData = await partiesRes.json();
-
-        setEncounters(encountersData || []);
-        setCharacters(charactersData || []);
-        setMonsterTemplates(monstersData || []);
-        setCombatState(combatData || null);
-        setParties(partiesData || []);
-      } catch (err) {
-        setError(err instanceof Error ? err.message : 'Failed to load data');
-      } finally {
-        setLoading(false);
-        setLoadingTemplates(false);
-      }
-    };
-
-    loadData();
-  }, []);
-
-  useEffect(() => {
-    const handleEscape = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        setSelectedDetailCombatantId(null);
-      }
-    };
-    if (selectedDetailCombatantId) {
-      document.addEventListener('keydown', handleEscape);
-      return () => document.removeEventListener('keydown', handleEscape);
-    }
-  }, [selectedDetailCombatantId]);
-
-  // Scroll to initiative panel when it opens
-  useEffect(() => {
-    if (initiativeEditId && initiativePanelRef.current) {
-      // Use scrollIntoView for more reliable scrolling
-      initiativePanelRef.current.scrollIntoView({
-        behavior: 'smooth',
-        block: 'start',
-      });
-    }
-  }, [initiativeEditId]);
-
-  // Auto-dismiss toast after 3 seconds
-  useEffect(() => {
-    if (toast) {
-      const timer = setTimeout(() => setToast(null), 3000);
-      return () => clearTimeout(timer);
-    }
-  }, [toast]);
-
-  // Keep ref in sync with setupCombatants state for duplicate detection
-  useEffect(() => {
-    setupCombatantsRef.current = setupCombatants;
-  }, [setupCombatants]);
-
-  const saveCombatState = async (state: CombatState | null) => {
-    const prev = combatState;
-    try {
-      setError(null);
-      setCombatState(state);
-      if (state) {
-        const response = await fetch('/api/combat', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(state),
-        });
-        if (!response.ok) throw new Error('Failed to save combat state');
-      }
-    } catch (err) {
-      setCombatState(prev);
-      setError(err instanceof Error ? err.message : 'Failed to save combat state');
-    }
-  };
-
-  const addCombatantToSetup = (combatant: CombatantState) => {
-    // Add to setupCombatants instead of starting combat
-    setSetupCombatants(prev => [...prev, combatant]);
-    // Keep ref in sync for immediate duplicate detection
-    setupCombatantsRef.current = [...setupCombatantsRef.current, combatant];
-    setShowQuickEntryType(null);
-  };
-
-  const removeCombatantFromSetup = (id: string) => {
-    setSetupCombatants(prev => prev.filter(c => c.id !== id));
-  };
-
-  const cancelLairForm = () => { setShowLairForm(false); setLairFormName(''); setLairFormSeedMonster(''); };
-
-  const confirmAddLair = () => {
-    const name = lairFormName.trim();
-    if (!name) return;
-    if (!combatState) {
-      // Setup phase
-      const lair = buildLairCombatant(name, lairFormSeedMonster, setupCombatants);
-      setSetupCombatants(prev => [...prev, lair]);
-      setupCombatantsRef.current = [...setupCombatantsRef.current, lair];
-    } else {
-      // Active combat
-      const lair = buildLairCombatant(name, lairFormSeedMonster, combatState.combatants);
-      const currentCombatantId = combatState.combatants[combatState.currentTurnIndex]?.id;
-      const updatedList = sortCombatants([...combatState.combatants, lair]);
-      const newTurnIndex = currentCombatantId
-        ? updatedList.findIndex(c => c.id === currentCombatantId)
-        : combatState.currentTurnIndex;
-      saveCombatState({
-        ...combatState,
-        combatants: updatedList,
-        currentTurnIndex: newTurnIndex !== -1 ? newTurnIndex : 0,
-      });
-    }
-    cancelLairForm();
-  };
-
-  const selectParty = (partyId: string | null) => {
-    setSelectedPartyId(partyId);
-  };
-
-  const addCombatantToActiveSession = (combatant: CombatantState) => {
-    // Add combatant directly to active combat, re-sorting if initiative has been rolled.
-    // Note: New combatants are added with initiative: 0. If initiative has been rolled,
-    // they will appear at the bottom of the turn order. Users can manually adjust
-    // initiative via the Manual Entry option if needed.
-    if (!combatState) return;
-
-    // Handle renumbering if there are duplicates with the same base name
-    const cleanName = combatant.name.replace(/\s+\d+$/, '');
-    
-    // Find existing combatants with the same base name
-    const existingWithSameName = combatState.combatants.filter(c => {
-      const cleanExisting = c.name.replace(/\s+\d+$/, '');
-      return cleanExisting === cleanName;
-    });
-
-    let finalCombatant = combatant;
-    let updatedCombatantsList = combatState.combatants;
-
-    if (existingWithSameName.length > 0) {
-      // There are duplicates - need to renumber all of them including the new one
-      // Renumber the existing ones sequentially
-      const updatedCombatants = combatState.combatants.map(c => {
-        const cleanExisting = c.name.replace(/\s+\d+$/, '');
-        if (cleanExisting === cleanName) {
-          // Find which number this should be in the sequence
-          const indexInDuplicates = existingWithSameName.findIndex(dup => dup.id === c.id);
-          return { ...c, name: `${cleanName} ${indexInDuplicates + 1}` };
-        }
-        return c;
-      });
-      
-      // Add the new combatant with the next number
-      const newCombatantIndex = existingWithSameName.length + 1;
-      finalCombatant = { ...combatant, name: `${cleanName} ${newCombatantIndex}` };
-      
-      updatedCombatantsList = [...updatedCombatants, finalCombatant];
-    } else {
-      // No duplicates, just add it
-      updatedCombatantsList = [...combatState.combatants, finalCombatant];
-    }
-
-    // Track the current combatant's ID to maintain turn pointer
-    const currentCombatantId = combatState.combatants[combatState.currentTurnIndex]?.id;
-    
-    // Only sort if initiative has been rolled; otherwise just append
-    const sortedCombatants = hasInitiativeBeenRolled()
-      ? sortCombatants(updatedCombatantsList)
-      : updatedCombatantsList;
-
-    // Find the index of the current combatant in the new list to preserve turn continuity
-    const newTurnIndex = currentCombatantId
-      ? sortedCombatants.findIndex(c => c.id === currentCombatantId)
-      : combatState.currentTurnIndex;
-
-    saveCombatState({
-      ...combatState,
-      combatants: sortedCombatants,
-      currentTurnIndex: newTurnIndex !== -1 ? newTurnIndex : 0,
-    });
-    setShowQuickEntryType(null);
-    setShowCombatantModal(false);
-  };
-
-  const addCombatantFromLibrary = (
-    item: Monster | Character,
-    type: 'monster' | 'player',
-    idPrefix: string
-  ) => {
-    try {
-      console.log('Adding combatant:', item.name);
-
-      const combatant: CombatantState = buildCombatantFromSource(item, type, idPrefix);
-
-      if (!combatState) {
-        // During setup phase - handle renumbering if needed
-        const cleanName = combatant.name.replace(/\s+\d+$/, '');
-        
-        // Find existing combatants with the same base name
-        const existingWithSameName = setupCombatantsRef.current.filter(c => {
-          const cleanExisting = c.name.replace(/\s+\d+$/, '');
-          return cleanExisting === cleanName;
-        });
-
-        if (existingWithSameName.length > 0) {
-          // There are duplicates - need to renumber all of them including the new one
-          // Renumber the existing ones sequentially
-          const updatedCombatants = setupCombatantsRef.current.map(c => {
-            const cleanExisting = c.name.replace(/\s+\d+$/, '');
-            if (cleanExisting === cleanName) {
-              // Find which number this should be in the sequence
-              const indexInDuplicates = existingWithSameName.findIndex(dup => dup.id === c.id);
-              return { ...c, name: `${cleanName} ${indexInDuplicates + 1}` };
-            }
-            return c;
-          });
-          
-          // Add the new combatant with the next number
-          const newCombatantIndex = existingWithSameName.length + 1;
-          combatant.name = `${cleanName} ${newCombatantIndex}`;
-          
-          const finalCombatants = [...updatedCombatants, combatant];
-          setSetupCombatants(finalCombatants);
-          setupCombatantsRef.current = finalCombatants;
-        } else {
-          // No duplicates, just add it
-          const newCombatants = [...setupCombatantsRef.current, combatant];
-          setSetupCombatants(newCombatants);
-          setupCombatantsRef.current = newCombatants;
-        }
-      } else {
-        // During active combat
-        addCombatantToActiveSession(combatant);
-      }
-
-      setShowCombatantModal(false);
-    } catch (err) {
-      const errMsg = err instanceof Error ? err.message : 'Failed to add combatant';
-      console.error('Error adding combatant:', err);
-      setError(errMsg);
-    }
-  };
-
-  const startCombatWithSetupCombatants = () => {
-    // Combine setup combatants with characters from library
-    const combatants: CombatantState[] = [...setupCombatants];
-
-    // Add characters: only party members if a party is selected, otherwise all characters
-    const charactersToAdd = resolveCharactersForCombat(selectedPartyId, parties, characters, setupCombatants);
-
-    charactersToAdd.forEach(character => {
-      const c = buildCombatantFromSource(character, 'player', 'character');
-      combatants.push({ ...c, id: `character-${character.id}` });
-    });
-
-    // Add monsters from selected encounter if any
-    if (selectedEncounterId) {
-      const encounter = encounters.find(e => e.id === selectedEncounterId);
-      if (encounter) {
-        encounter.monsters.forEach((monster, idx) => {
-          const c = buildCombatantFromSource(monster, 'monster', 'monster');
-          combatants.push({ ...c, id: `monster-${monster.id}-${idx}`, name: `${monster.name} ${idx + 1}` });
-        });
-      }
-    }
-
-    const encounter = selectedEncounterId ? encounters.find(e => e.id === selectedEncounterId) : undefined;
-
-    const newState: CombatState = {
-      id: crypto.randomUUID(),
-      userId: '',
-      encounterId: selectedEncounterId || undefined,
-      encounterDescription: encounter?.description,
-      combatants,
-      currentRound: 1,
-      currentTurnIndex: 0,
-      isActive: true,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-    };
-
-    saveCombatState(newState);
-    setSetupCombatants([]);
-  };
-
-  const startCombat = () => {
-    // Starts combat including both library encounter combatants and any accumulated setup combatants.
-    startCombatWithSetupCombatants();
-  };
-
-  const endCombat = () => {
-    if (confirm('Are you sure you want to end combat?')) {
-      if (combatState) clearCombatHistory(combatState.id);
-      saveCombatState(null);
-      setSetupCombatants([]);
-      setSelectedPartyId(null);
-    }
-  };
-
-  const restartRound = () => {
-    if (!combatState) return;
-    saveCombatState({
-      ...combatState,
-      currentTurnIndex: 0,
-    });
-  };
-
-  const rollInitiative = () => {
-    if (!combatState) return;
-
-    // Roll initiative for all non-lair combatants (lair slots are always initiative 20)
-    const updatedCombatants = combatState.combatants.map(c => {
-      if (c.type === 'lair') return c;
-      const initiativeRoll = buildInitiativeRoll(c);
-      return { ...c, initiative: initiativeRoll.total, initiativeRoll };
-    });
-
-    saveCombatState({
-      ...combatState,
-      combatants: sortCombatants(updatedCombatants),
-      currentTurnIndex: 0,
-    });
-    setInitiativeMode(false);
-  };
-
-  const nextTurn = () => {
-    if (!combatState) return;
-
-    let nextIndex = combatState.currentTurnIndex + 1;
-    let nextRound = combatState.currentRound;
-    let baseCombatants = combatState.combatants;
-
-    if (nextIndex >= combatState.combatants.length) {
-      nextIndex = 0;
-      nextRound += 1;
-
-      // Decrement condition durations, collect expiring conditions, and remove expired ones (single pass)
-      const { updatedCombatants, expiring } = processRoundEnd(combatState.combatants);
-      if (expiring.length > 0) {
-        const lines = expiring.map(e => `• ${e.combatantName}: ${e.conditionName}`).join('\n');
-        alert(`Conditions expired:\n${lines}`);
-      }
-      baseCombatants = updatedCombatants;
-    }
-
-    // Reset legendary action pool for the incoming combatant (both mid-round and round-end paths)
-    const combatants = resetIncomingLegendaryPool(baseCombatants, nextIndex);
-
-    saveCombatState({
-      ...combatState,
-      combatants,
-      currentTurnIndex: nextIndex,
-      currentRound: nextRound,
-    });
-  };
-
-  const updateCombatant = (id: string, updates: Partial<CombatantState>) => {
-    if (!combatState) return;
-
-    const updatedCombatants = combatState.combatants.map(c =>
-      c.id === id ? { ...c, ...updates } : c
-    );
-
-    saveCombatState({
-      ...combatState,
-      combatants: updatedCombatants,
-    });
-  };
-
-  const updateCombatantInitiativeSettings = (id: string, adv: boolean, fb: number) => {
-    updateCombatant(id, { initiativeAdvantage: adv, initiativeFlatBonus: fb });
-  };
-
-  const removeCombatant = (id: string) => {
-    if (!combatState) return;
-
-    const updatedCombatants = combatState.combatants.filter(c => c.id !== id);
-    saveCombatState({
-      ...combatState,
-      combatants: updatedCombatants,
-    });
-  };
-
-  const setInitiativeRoll = (combatantId: string, initiativeRoll: InitiativeRoll) => {
-    if (!combatState) return;
-
-    // Track the current combatant's ID to maintain turn pointer
-    const currentCombatantId = combatState.combatants[combatState.currentTurnIndex]?.id;
-
-    const updatedCombatants = combatState.combatants.map(c =>
-      c.id === combatantId 
-        ? { ...c, initiative: initiativeRoll.total, initiativeRoll }
-        : c
-    );
-
-    const sortedCombatants = sortCombatants(updatedCombatants);
-
-    // Find the index of the current combatant in the new sorted list to preserve turn continuity
-    const newTurnIndex = currentCombatantId
-      ? sortedCombatants.findIndex(c => c.id === currentCombatantId)
-      : 0;
-
-    saveCombatState({
-      ...combatState,
-      combatants: sortedCombatants,
-      currentTurnIndex: newTurnIndex !== -1 ? newTurnIndex : 0,
-    });
-  };
-
-  const hasInitiativeBeenRolled = () => {
-    return combatState?.combatants.some(c => c.initiativeRoll);
-  };
-
-  const getDisplayCombatants = () => {
-    if (!combatState) return [];
-
-    if (hasInitiativeBeenRolled()) {
-      return sortCombatants([...combatState.combatants]);
-    } else {
-      // Before initiative, group players at top, monsters at bottom
-      const players = combatState.combatants.filter(c => c.type === 'player');
-      const monsters = combatState.combatants.filter(c => c.type === 'monster');
-      return [...players, ...monsters];
-    }
-  };
-
-  // Combatants that still need initiative (initiative === 0) sorted alphabetically by name
-  const zeroInitiative = useMemo(() => {
-    if (!combatState) return [] as CombatantState[];
-    return [...combatState.combatants.filter(c => c.initiative === 0)].sort((a, b) => a.name.localeCompare(b.name));
-  }, [combatState]);
-
-  // Filtered view for zero-initiative list (all/player/monster)
-  const filteredZeroInitiative = zeroInitiative.filter(c => initiativeFilter === 'all' || c.type === initiativeFilter);
-
-  if (loading) {
-    return (
-      <div className="min-h-screen bg-gray-900 text-white flex items-center justify-center">
-        <div className="text-center">
-          <p className="text-gray-400">Loading combat data...</p>
-        </div>
-      </div>
-    );
+  if (combat.loading) return <div className="min-h-screen bg-gray-900 flex items-center justify-center text-white">Loading combat...</div>;
+  if (!combat.combatState) {
+    return <CombatSetupView combat={combat} user={user} />;
   }
-
-  if (!combatState) {
-    // Setup phase: show option to start new combat with library combatants or quick entry
-    return (
-      <div className="min-h-screen bg-gray-900 text-white">
-        <div className="container mx-auto px-4 py-8">
-          <div className="flex justify-between items-center mb-8">
-            <h1 className="text-3xl font-bold">Combat Tracker</h1>
-            <Link href="/" className="bg-gray-700 hover:bg-gray-600 px-4 py-2 rounded">
-              Back to Home
-            </Link>
-          </div>
-
-          {error && (
-            <div className="p-4 bg-red-900 border border-red-700 rounded text-red-200 mb-6">
-              {error}
-            </div>
-          )}
-
-          <div className="bg-gray-800 rounded-lg p-6 max-w-4xl mx-auto">
-            <h2 className="text-2xl font-bold mb-6">Start New Combat</h2>
-            
-            <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-              {/* Option 1: Use existing characters */}
-              <div className="border border-gray-700 rounded-lg p-4">
-                <h3 className="text-lg font-semibold mb-3">From Library</h3>
-                <p className="text-gray-400 text-sm mb-4">
-                  Use your saved characters and encounters
-                </p>
-                
-                <div className="mb-4">
-                  <label className="block text-sm mb-2">Select Encounter (Optional)</label>
-                  <select
-                    value={selectedEncounterId}
-                    onChange={(e) => setSelectedEncounterId(e.target.value)}
-                    className="w-full bg-gray-700 rounded px-3 py-2 text-white text-sm"
-                  >
-                    <option value="">No encounter</option>
-                    {encounters.map(encounter => (
-                      <option key={encounter.id} value={encounter.id}>
-                        {encounter.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-
-                <div className="mb-4">
-                  <label className="block text-sm mb-2">Select Party (Optional)</label>
-                  <select
-                    value={selectedPartyId ?? ''}
-                    onChange={(e) => selectParty(e.target.value || null)}
-                    className="w-full bg-gray-700 rounded px-3 py-2 text-white text-sm"
-                  >
-                    <option value="">No party (all characters)</option>
-                    {parties.map(party => (
-                      <option key={party.id} value={party.id}>
-                        {party.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-
-                <div className="mb-4">
-                  <p className="text-gray-400 text-xs">
-                    Characters: {resolveCharactersForCombat(selectedPartyId, parties, characters, setupCombatants).length} |
-                    Monsters: {selectedEncounterId ? encounters.find(e => e.id === selectedEncounterId)?.monsters.length || 0 : 0}
-                  </p>
-                </div>
-
-                <button
-                  onClick={startCombat}
-                  disabled={characters.length === 0}
-                  className="w-full bg-green-600 hover:bg-green-700 disabled:bg-gray-600 px-4 py-2 rounded text-sm font-semibold transition-colors"
-                >
-                  Start Combat
-                </button>
-                {characters.length === 0 && (
-                  <p className="text-red-400 text-xs mt-2">Need at least one character</p>
-                )}
-              </div>
-
-              {/* Option 2: Quick entry */}
-              <div className="border border-blue-700 border-2 rounded-lg p-4 bg-blue-900 bg-opacity-20">
-                <h3 className="text-lg font-semibold mb-3 text-blue-300">Quick Entry</h3>
-                <p className="text-gray-400 text-sm mb-4">
-                  Add combatants on the fly during session
-                </p>
-                
-                <p className="text-xs text-gray-400 mb-4">
-                  Enter name, dexterity, HP, and optional initiative. Perfect for DMs managing player characters elsewhere.
-                </p>
-
-                {setupCombatants.length > 0 && (
-                  <div className="mb-4 bg-gray-800 rounded p-3">
-                    <p className="text-xs text-gray-400 mb-2">Quick Entry Combatants:</p>
-                    <div className="space-y-2">
-                      {setupCombatants.map(combatant => (
-                        <div key={combatant.id} className="flex justify-between items-center bg-gray-700 rounded px-2 py-1 text-sm">
-                          <span className="text-white">
-                            {combatant.name} 
-                            <span className={`ml-2 px-2 py-0.5 rounded text-xs ${combatant.type === 'player' ? 'bg-blue-600' : 'bg-red-600'}`}>
-                              {combatant.type}
-                            </span>
-                          </span>
-                          <button
-                            onClick={() => removeCombatantFromSetup(combatant.id)}
-                            className="text-red-400 hover:text-red-300 text-xs"
-                            aria-label={`Remove ${combatant.name}`}
-                          >
-                            Remove
-                          </button>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                <div className="space-y-2">
-                  <button
-                    onClick={() => setShowCombatantModal(true)}
-                    className="w-full bg-blue-600 hover:bg-blue-700 px-4 py-2 rounded text-sm font-semibold transition-colors"
-                  >
-                    + Add Party Member
-                  </button>
-                  <button
-                    onClick={() => setShowCombatantModal(true)}
-                    className="w-full bg-red-600 hover:bg-red-700 px-4 py-2 rounded text-sm font-semibold transition-colors"
-                  >
-                    + Add Enemy
-                  </button>
-                  <button
-                    onClick={() => setShowLairForm(true)}
-                    className="w-full bg-purple-600 hover:bg-purple-700 px-4 py-2 rounded text-sm font-semibold transition-colors"
-                  >
-                    + Add Lair
-                  </button>
-                  {setupCombatants.length > 0 && (
-                    <button
-                      onClick={startCombatWithSetupCombatants}
-                      data-testid="start-combat-quick"
-                      className="w-full bg-green-600 hover:bg-green-700 px-4 py-2 rounded text-sm font-semibold transition-colors mt-2"
-                    >
-                      Start Combat
-                    </button>
-                  )}
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {showCombatantModal && (
-            <QuickCombatantModal
-              onAddMonster={(monster) => addCombatantFromLibrary(monster, 'monster', 'monster')}
-              onAddCharacter={(character) => addCombatantFromLibrary(character, 'player', 'character')}
-              onClose={() => setShowCombatantModal(false)}
-              monsterTemplates={monsterTemplates}
-              characterTemplates={characters}
-              loadingTemplates={loadingTemplates}
-              userId={user?.userId}
-            />
-          )}
-
-          {showLairForm && (
-            <LairForm
-              seedOptions={setupCombatants.filter(c => c.type !== 'lair' && (c.lairActions ?? []).length > 0).map(c => c.name)}
-              lairName={lairFormName}
-              seedMonster={lairFormSeedMonster}
-              onNameChange={setLairFormName}
-              onSeedChange={setLairFormSeedMonster}
-              onConfirm={confirmAddLair}
-              onCancel={cancelLairForm}
-            />
-          )}
-        </div>
-      </div>
-    );
-  }
-
-  const currentCombatant = combatState.combatants[combatState.currentTurnIndex];
-
-  // Show initiative entry modal if in initiative mode
-  if (initiativeMode && combatState) {
-    return (
-      <div className="min-h-screen bg-gray-900 text-white">
-        <div className="container mx-auto px-4 py-8">
-          <div className="flex justify-between items-center mb-8">
-            <h1 className="text-3xl font-bold">Initiative</h1>
-            <Link href="/" className="bg-gray-700 hover:bg-gray-600 px-4 py-2 rounded">
-              Back to Home
-            </Link>
-          </div>
-
-          {error && (
-            <div className="p-4 bg-red-900 border border-red-700 rounded text-red-200 mb-6">
-              {error}
-            </div>
-          )}
-
-          <div className="space-y-4 max-w-2xl mx-auto">
-            {combatState.combatants.map((combatant) => (
-              <InitiativeEntry
-                key={combatant.id}
-                combatant={combatant}
-                onSet={(initiativeRoll) => setInitiativeRoll(combatant.id, initiativeRoll)}
-                onSettingsChange={(adv, fb) => updateCombatantInitiativeSettings(combatant.id, adv, fb)}
-              />
-            ))}
-          </div>
-
-          <div className="flex gap-2 justify-center mt-8">
-            <button
-              onClick={() => setInitiativeMode(false)}
-              className="bg-green-600 hover:bg-green-700 px-6 py-3 rounded text-lg font-semibold"
-            >
-              Start Combat
-            </button>
-            <button
-              onClick={() => setInitiativeMode(false)}
-              className="bg-gray-600 hover:bg-gray-700 px-6 py-3 rounded text-lg font-semibold"
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="min-h-screen bg-gray-900 text-white" data-testid="combat-screen">
-      <div className="container mx-auto px-4 py-8">
-        <div className="flex justify-between items-center mb-6">
-          <div>
-            <div className="flex items-center gap-2">
-              <h1 className="text-3xl font-bold">Combat Tracker</h1>
-              {combatState.encounterDescription && (
-                <button
-                  onClick={() => setShowEncounterDescription(true)}
-                  className="hover:opacity-80 transition-opacity"
-                  title="See Encounter Description"
-                  type="button"
-                >
-                  <svg
-                    className="w-6 h-6 text-gray-400 hover:text-gray-300 cursor-pointer"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path
-                      fillRule="evenodd"
-                      d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7-4a1 1 0 11-2 0 1 1 0 012 0zM9 9a1 1 0 000 2v3a1 1 0 001 1h1a1 1 0 100-2v-3a1 1 0 00-1-1H9z"
-                      clipRule="evenodd"
-                    />
-                  </svg>
-                </button>
-              )}
-            </div>
-            <div className="flex items-center gap-2">
-              <p className="text-gray-400">Round {combatState.currentRound}</p>
-              <CombatInfoIcon combatants={combatState.combatants} />
-            </div>
-          </div>
-          <div className="flex gap-2 flex-wrap">
-            <button
-              onClick={() => setShowCombatantModal(true)}
-              className="bg-blue-500 hover:bg-blue-600 px-4 py-2 rounded text-sm"
-              title="Add a party member mid-combat"
-            >
-              + Add Party Member
-            </button>
-            <button
-              onClick={() => setShowCombatantModal(true)}
-              className="bg-red-500 hover:bg-red-600 px-4 py-2 rounded text-sm"
-              title="Add an enemy mid-combat"
-            >
-              + Add Enemy
-            </button>
-            <button
-              onClick={() => setShowLairForm(true)}
-              className="bg-purple-600 hover:bg-purple-700 px-4 py-2 rounded text-sm"
-              title="Add a lair actions slot"
-            >
-              + Add Lair
-            </button>
-            <button
-              onClick={restartRound}
-              className="bg-blue-700 hover:bg-blue-800 px-4 py-2 rounded text-sm"
-              title="Reset turn order to the first combatant"
-            >
-              Restart Round
-            </button>
-            <button
-              onClick={endCombat}
-              className="bg-red-600 hover:bg-red-700 px-4 py-2 rounded"
-            >
-              End Combat
-            </button>
-            <Link href="/" className="bg-gray-700 hover:bg-gray-600 px-4 py-2 rounded">
-              Back to Home
-            </Link>
-          </div>
-        </div>
-
-        {error && (
-          <div className="p-4 bg-red-900 border border-red-700 rounded text-red-200 mb-6">
-            {error}
-          </div>
-        )}
-
-        {/* Initiative Entry Blocks for Combatants with 0 Initiative */}
-        {zeroInitiative.length > 0 && (
-          <div ref={initiativePanelRef} className="mb-6 space-y-4">
-            {/* Filter controls - only show when there are zero-initiative combatants */}
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <p className="text-sm text-gray-400">Show:</p>
-                <div className="flex gap-2">
-                  <button
-                    onClick={() => setInitiativeFilter('all')}
-                    className={`px-3 py-1 rounded text-sm ${initiativeFilter === 'all' ? 'bg-gray-700 text-white' : 'bg-gray-800 hover:bg-gray-700 text-gray-300'}`}
-                  >
-                    All
-                  </button>
-                  <button
-                    onClick={() => setInitiativeFilter('player')}
-                    className={`px-3 py-1 rounded text-sm ${initiativeFilter === 'player' ? 'bg-gray-700 text-white' : 'bg-gray-800 hover:bg-gray-700 text-gray-300'}`}
-                  >
-                    Players
-                  </button>
-                  <button
-                    onClick={() => setInitiativeFilter('monster')}
-                    className={`px-3 py-1 rounded text-sm ${initiativeFilter === 'monster' ? 'bg-gray-700 text-white' : 'bg-gray-800 hover:bg-gray-700 text-gray-300'}`}
-                  >
-                    Monsters
-                  </button>
-                </div>
-              </div>
-
-              <p className="text-sm text-gray-400">{zeroInitiative.length} need initiative</p>
-            </div>
-
-            {filteredZeroInitiative.length === 0 ? (
-              <div className="p-4 bg-gray-800 rounded text-gray-400">No combatants match the selected filter.</div>
-            ) : (
-              filteredZeroInitiative.map((combatant) => (
-                <div key={combatant.id} className="bg-gray-800 rounded-lg p-6 border border-gray-700">
-                  <InitiativeEntry
-                    combatant={combatant}
-                    onSet={(initiativeRoll) => {
-                      setInitiativeRoll(combatant.id, initiativeRoll);
-                    }}
-                    onSettingsChange={(adv, fb) => updateCombatantInitiativeSettings(combatant.id, adv, fb)}
-                  />
-                </div>
-              ))
-            )}
-          </div>
-        )}
-
-        {/* Initiative Entry Panel - for clicking on individual combatants */}
-        {initiativeEditId && combatState && !combatState.combatants.some(c => c.initiative === 0) && (
-          <div className="mb-6 bg-gray-800 rounded-lg p-6 border border-gray-700">
-            <InitiativeEntry
-              key={initiativeEditId}
-              combatant={combatState.combatants.find(c => c.id === initiativeEditId)!}
-              onSet={(initiativeRoll) => {
-                setInitiativeRoll(initiativeEditId, initiativeRoll);
-                setInitiativeEditId(null);
-              }}
-              onClose={() => setInitiativeEditId(null)}
-              onSettingsChange={(adv, fb) => updateCombatantInitiativeSettings(initiativeEditId, adv, fb)}
-            />
-          </div>
-        )}
-
-        {hasInitiativeBeenRolled() ? (
-          // Display sorted by initiative
-          <div className="space-y-2" data-testid="initiative-order">
-            <h2 className="text-xl font-semibold text-yellow-400 mb-4">Initiative Order</h2>
-            {getDisplayCombatants().map((combatant, idx) => {
-              // Find the actual index in combatState for isActive check
-              const actualIdx = combatState.combatants.findIndex(c => c.id === combatant.id);
-              const isActive = actualIdx === combatState.currentTurnIndex;
-              if (combatant.type === 'lair') {
-                return (
-                  <div key={combatant.id} data-testid={isActive ? 'lair-active' : 'lair-slot-badge'}>
-                    <LairActionsSlot
-                      combatant={combatant}
-                      isActive={isActive}
-                      onUpdate={(updates) => updateCombatant(combatant.id, updates)}
-                      onNextTurn={nextTurn}
-                    />
-                    {!isActive && (
-                      <button
-                        type="button"
-                        data-testid="lair-slot-remove"
-                        className="text-xs text-red-400 hover:text-red-300 mt-1 ml-2"
-                        onClick={(e) => {
-                          const rect = (e.currentTarget as HTMLButtonElement).getBoundingClientRect();
-                          setRemoveConfirmId(combatant.id);
-                          setRemoveConfirmPosition({ top: rect.bottom + window.scrollY, left: rect.left + window.scrollX });
-                        }}
-                      >
-                        Remove
-                      </button>
-                    )}
-                  </div>
-                );
-              }
-              return (
-                <CombatantCard
-                  key={combatant.id}
-                  combatId={combatState.id}
-                  combatant={combatant}
-                  isActive={isActive}
-                  onUpdate={(updates) => updateCombatant(combatant.id, updates)}
-                  onRemove={() => removeCombatant(combatant.id)}
-                  onNextTurn={nextTurn}
-                  onShowDetails={(id, pos) => {
-                    setSelectedDetailCombatantId(id);
-                    setDetailPosition(pos);
-                  }}
-                  onSetInitiative={setInitiativeEditId}
-                  onShowRemoveConfirm={(id, pos) => {
-                    setRemoveConfirmId(id);
-                    setRemoveConfirmPosition(pos);
-                  }}
-                  allCombatants={combatState.combatants}
-                  onUpdateCombatant={(id, updates) => updateCombatant(id, updates)}
-                />
-              );
-            })}
-          </div>
-        ) : (
-          // Display grouped by type
-          <div className="space-y-6" data-testid="combatants-list">
-            <div>
-              <h2 className="text-xl font-semibold text-blue-400 mb-3">Party</h2>
-              <div className="space-y-2">
-                {getDisplayCombatants()
-                  .filter(c => c.type === 'player')
-                  .map((combatant, idx) => {
-                    const actualIdx = combatState.combatants.findIndex(c => c.id === combatant.id);
-                    return (
-                      <CombatantCard
-                        key={combatant.id}
-                        combatId={combatState.id}
-                        combatant={combatant}
-                        isActive={actualIdx === combatState.currentTurnIndex}
-                        onUpdate={(updates) => updateCombatant(combatant.id, updates)}
-                        onRemove={() => removeCombatant(combatant.id)}
-                        onNextTurn={nextTurn}
-                        onShowDetails={(id, pos) => {
-                    setSelectedDetailCombatantId(id);
-                    setDetailPosition(pos);
-                  }}
-                        onSetInitiative={setInitiativeEditId}
-                        onShowRemoveConfirm={(id, pos) => {
-                          setRemoveConfirmId(id);
-                          setRemoveConfirmPosition(pos);
-                        }}
-                        allCombatants={combatState.combatants}
-                        onUpdateCombatant={(id, updates) => updateCombatant(id, updates)}
-                      />
-                    );
-                  })}
-              </div>
-            </div>
-
-            <div>
-              <h2 className="text-xl font-semibold text-red-400 mb-3">Enemies</h2>
-              <div className="space-y-2">
-                {getDisplayCombatants()
-                  .filter(c => c.type === 'monster')
-                  .map((combatant, idx) => {
-                    const actualIdx = combatState.combatants.findIndex(c => c.id === combatant.id);
-                    return (
-                      <CombatantCard
-                        key={combatant.id}
-                        combatId={combatState.id}
-                        combatant={combatant}
-                        isActive={actualIdx === combatState.currentTurnIndex}
-                        onUpdate={(updates) => updateCombatant(combatant.id, updates)}
-                        onRemove={() => removeCombatant(combatant.id)}
-                        onNextTurn={nextTurn}
-                        onShowDetails={(id, pos) => {
-                    setSelectedDetailCombatantId(id);
-                    setDetailPosition(pos);
-                  }}
-                        onSetInitiative={setInitiativeEditId}
-                        onShowRemoveConfirm={(id, pos) => {
-                          setRemoveConfirmId(id);
-                          setRemoveConfirmPosition(pos);
-                        }}
-                        allCombatants={combatState.combatants}
-                        onUpdateCombatant={(id, updates) => updateCombatant(id, updates)}
-                      />
-                    );
-                  })}
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* Combatant Modal for both library and custom */}
-        {showCombatantModal && (
-          <QuickCombatantModal
-            onAddMonster={(monster) => addCombatantFromLibrary(monster, 'monster', 'monster')}
-            onAddCharacter={(character) => addCombatantFromLibrary(character, 'player', 'character')}
-            onClose={() => setShowCombatantModal(false)}
-            monsterTemplates={monsterTemplates}
-            characterTemplates={characters}
-            loadingTemplates={loadingTemplates}
-            userId={user?.userId}
-          />
-        )}
-
-        {/* Add Lair Form */}
-        {showLairForm && combatState && (
-          <LairForm
-            seedOptions={combatState.combatants.filter(c => c.type !== 'lair' && (c.lairActions ?? []).length > 0).map(c => c.name)}
-            lairName={lairFormName}
-            seedMonster={lairFormSeedMonster}
-            onNameChange={setLairFormName}
-            onSeedChange={setLairFormSeedMonster}
-            onConfirm={confirmAddLair}
-            onCancel={cancelLairForm}
-          />
-        )}
-
-        {/* Detail Popup */}
-        {selectedDetailCombatantId && combatState && detailPosition && (
-          <div 
-            className="absolute bg-gray-800 rounded-lg p-6 max-w-2xl w-full max-h-[85vh] overflow-y-auto shadow-xl border border-gray-700 z-50"
-            style={{
-              top: '10px',
-              left: `${detailPosition.left}px`,
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            {(() => {
-              const combatant = combatState.combatants.find(c => c.id === selectedDetailCombatantId);
-              if (!combatant) return <div>Combatant not found</div>;
-              
-              return (
-                <>
-                  <div className="flex justify-between items-start gap-4 mb-4">
-                    <div className="flex items-center gap-2 flex-1">
-                      <h2 className="text-2xl font-bold">
-                        {combatant.name}
-                      </h2>
-                      <span className={`px-3 py-1 rounded-full text-xs font-semibold ${combatant.type === 'player' ? 'bg-blue-600 text-blue-100' : 'bg-red-600 text-red-100'}`}>
-                        {combatant.type === 'player' ? 'Character' : 'Monster'}
-                      </span>
-                    </div>
-                    <button
-                      onClick={() => setSelectedDetailCombatantId(null)}
-                      className="text-gray-400 hover:text-gray-200 text-2xl flex-shrink-0"
-                    >
-                      ×
-                    </button>
-                  </div>
-
-                  <div className="space-y-4">
-                    <div className="grid grid-cols-3 gap-4">
-                      <div>
-                        <p className="text-gray-400 text-sm">HP</p>
-                        <p className="text-lg font-bold">{combatant.hp} / {combatant.maxHp}</p>
-                      </div>
-                      <div>
-                        <p className="text-gray-400 text-sm">AC</p>
-                        <p className="text-lg font-bold">{combatant.ac}</p>
-                      </div>
-                      <div>
-                        <p className="text-gray-400 text-sm">Initiative</p>
-                        <p className="text-lg font-bold">{combatant.initiative}</p>
-                      </div>
-                    </div>
-
-                    {combatant.actions && combatant.actions.length > 0 && (
-                      <div>
-                        <p className="text-gray-400 text-sm mb-2 font-semibold">Actions</p>
-                        <div className="space-y-2">
-                          {combatant.actions.map((action) => (
-                            <div key={action.name} className="text-xs">
-                              <p className="font-bold text-white">{action.name}</p>
-                              <p className="text-gray-300">{action.description}</p>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-
-                    {combatant.bonusActions && combatant.bonusActions.length > 0 && (
-                      <div>
-                        <p className="text-gray-400 text-sm mb-2 font-semibold">Bonus Actions</p>
-                        <div className="space-y-2">
-                          {combatant.bonusActions.map((action) => (
-                            <div key={action.name} className="text-xs">
-                              <p className="font-bold text-white">{action.name}</p>
-                              <p className="text-gray-300">{action.description}</p>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-
-                    {combatant.reactions && combatant.reactions.length > 0 && (
-                      <div>
-                        <p className="text-gray-400 text-sm mb-2 font-semibold">Reactions</p>
-                        <div className="space-y-2">
-                          {combatant.reactions.map((action) => (
-                            <div key={action.name} className="text-xs">
-                              <p className="font-bold text-white">{action.name}</p>
-                              <p className="text-gray-300">{action.description}</p>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-
-                    <LegendaryActionsPanel
-                      combatant={combatant}
-                      onUpdate={(updates) => updateCombatant(combatant.id, updates)}
-                    />
-
-                    {combatant.lairActions && combatant.lairActions.length > 0 && (
-                      <div>
-                        <p className="text-gray-400 text-sm mb-2 font-semibold">Lair Actions</p>
-                        <div className="space-y-2">
-                          {combatant.lairActions.map((action) => (
-                            <div key={action.name} className="text-xs">
-                              <p className="font-bold text-white">{action.name}</p>
-                              <p className="text-gray-300">{action.description}</p>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-
-                    {combatant.abilityScores && (
-                      <div>
-                        <p className="text-gray-400 text-sm mb-2">Ability Scores</p>
-                        <div className="grid grid-cols-6 gap-2">
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">STR</p>
-                            <p className="font-bold">{combatant.abilityScores?.strength || 10}</p>
-                          </div>
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">DEX</p>
-                            <p className="font-bold">{combatant.abilityScores?.dexterity || 10}</p>
-                          </div>
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">CON</p>
-                            <p className="font-bold">{combatant.abilityScores?.constitution || 10}</p>
-                          </div>
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">INT</p>
-                            <p className="font-bold">{combatant.abilityScores?.intelligence || 10}</p>
-                          </div>
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">WIS</p>
-                            <p className="font-bold">{combatant.abilityScores?.wisdom || 10}</p>
-                          </div>
-                          <div className="bg-gray-700 p-2 rounded text-center">
-                            <p className="text-xs text-gray-400">CHA</p>
-                            <p className="font-bold">{combatant.abilityScores?.charisma || 10}</p>
-                          </div>
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                </>
-              );
-            })()}
-          </div>
-        )}
-
-        {/* Encounter Description Modal */}
-        {showEncounterDescription && combatState?.encounterDescription && (
-          <div 
-            className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50"
-            onClick={() => setShowEncounterDescription(false)}
-          >
-            <div 
-              className="bg-gray-800 rounded-lg p-6 max-w-2xl w-full mx-4 max-h-[85vh] overflow-y-auto shadow-xl border border-gray-700"
-              onClick={(e) => e.stopPropagation()}
-            >
-              <div className="flex justify-between items-start gap-4 mb-4">
-                <h2 className="text-2xl font-bold">Encounter Description</h2>
-                <button
-                  onClick={() => setShowEncounterDescription(false)}
-                  className="text-gray-400 hover:text-gray-200 text-2xl flex-shrink-0"
-                >
-                  ×
-                </button>
-              </div>
-              <div className="text-gray-300 whitespace-pre-wrap leading-relaxed">
-                {combatState.encounterDescription}
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* Remove Confirmation Modal */}
-        {removeConfirmId && combatState && removeConfirmPosition && (
-          <div 
-            className="absolute bg-gray-800 rounded-lg p-6 max-w-sm w-80 shadow-xl border border-gray-700 z-50"
-            style={{
-              top: `${removeConfirmPosition.top}px`,
-              left: `${removeConfirmPosition.left}px`,
-            }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            {(() => {
-              const combatant = combatState.combatants.find(c => c.id === removeConfirmId);
-              if (!combatant) return null;
-              
-              return (
-                <>
-                  <p className="text-lg font-semibold mb-4">
-                    Remove <span className="text-red-400">{combatant.name}</span> from combat?
-                  </p>
-                  <div className="flex gap-3">
-                    <button
-                      onClick={() => {
-                        removeCombatant(removeConfirmId);
-                        setRemoveConfirmId(null);
-                        setRemoveConfirmPosition(null);
-                      }}
-                      data-testid="remove-confirm-button"
-                      className="flex-1 bg-red-600 hover:bg-red-700 px-4 py-2 rounded font-semibold"
-                    >
-                      Remove
-                    </button>
-                    <button
-                      onClick={() => {
-                        setRemoveConfirmId(null);
-                        setRemoveConfirmPosition(null);
-                      }}
-                      className="flex-1 bg-gray-700 hover:bg-gray-600 px-4 py-2 rounded font-semibold"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                </>
-              );
-            })()}
-          </div>
-        )}
-      </div>
-
-      {/* Toast Notification */}
-      {toast && (
-        <div
-          className={`fixed bottom-6 right-6 px-6 py-3 rounded-lg shadow-lg text-white font-semibold transition-opacity duration-300 z-50 ${
-            toast.type === 'success' ? 'bg-green-600' : 'bg-red-600'
-          }`}
-        >
-          {toast.message}
-        </div>
-      )}
-    </div>
-  );
+  return <ActiveCombatView combat={combat} user={user} />;
 }
-
-export function CombatantCard({
-  combatId,
-  combatant,
-  isActive,
-  onUpdate,
-  onRemove,
-  onNextTurn,
-  onShowDetails,
-  onSetInitiative,
-  onShowRemoveConfirm,
-  allCombatants,
-  onUpdateCombatant,
-}: {
-  combatId: string;
-  combatant: CombatantState;
-  isActive: boolean;
-  onUpdate: (updates: Partial<CombatantState>) => void;
-  onRemove: () => void;
-  onNextTurn?: () => void;
-  onShowDetails?: (combatantId: string, position: {top: number, left: number}) => void;
-  onSetInitiative?: (combatantId: string) => void;
-  onShowRemoveConfirm?: (combatantId: string, position: {top: number, left: number}) => void;
-  allCombatants?: CombatantState[];
-  onUpdateCombatant?: (combatantId: string, updates: Partial<CombatantState>) => void;
-}) {
-  const [isEditing, setIsEditing] = useState(false);
-  const [showConditions, setShowConditions] = useState(false);
-  const [hpAdjustment, setHpAdjustment] = useState('');
-  const [showTargeting, setShowTargeting] = useState(false);
-  const [selectedTargetId, setSelectedTargetId] = useState<string | null>(null);
-  const [damageInput, setDamageInput] = useState('');
-  const [newCondition, setNewCondition] = useState('');
-  const [conditionDuration, setConditionDuration] = useState('');
-  const [targetActionMode, setTargetActionMode] = useState<'damage' | 'condition' | null>(null);
-  const [hoveredTargetId, setHoveredTargetId] = useState<string | null>(null);
-  const [isTempMode, setIsTempMode] = useState(false);
-  const [selectedDamageType, setSelectedDamageType] = useState<DamageType | ''>('');
-  const [targetDamageType, setTargetDamageType] = useState<DamageType | ''>('');
-  const [showEffectsPanel, setShowEffectsPanel] = useState(false);
-  const [pendingPreset, setPendingPreset] = useState<{ label: string; kind: 'resistance' | 'immunity' | 'vulnerability'; choices: DamageType[] } | null>(null);
-  // Bumped after this card's own push/pop to keep the Undo button enabled state in sync
-  const [historyVersion, setHistoryVersion] = useState(0);
-
-  const adjustHp = (amount: number) => {
-    const prevHp = combatant.hp;
-    const prevTempHp = combatant.tempHp ?? 0;
-    if (amount < 0) {
-      const rawDamage = -amount;
-      let resultHp: number;
-      let resultTempHp: number;
-      if (selectedDamageType) {
-        const result = calcApplyDamageWithType(
-          prevHp,
-          prevTempHp,
-          rawDamage,
-          selectedDamageType,
-          {
-            damageResistances: combatant.damageResistances,
-            damageImmunities: combatant.damageImmunities,
-            damageVulnerabilities: combatant.damageVulnerabilities,
-            activeDamageEffects: combatant.activeDamageEffects,
-          },
-        );
-        resultHp = result.hp;
-        resultTempHp = result.tempHp;
-      } else {
-        const result = calcApplyDamage(prevHp, prevTempHp, rawDamage);
-        resultHp = result.hp;
-        resultTempHp = result.tempHp;
-      }
-      if (resultHp !== prevHp || resultTempHp !== prevTempHp) {
-        pushHpHistory(combatId, combatant.id, { hp: prevHp, tempHp: prevTempHp, type: 'damage', amount: rawDamage, timestamp: Date.now() });
-        setHistoryVersion(v => v + 1);
-      }
-      onUpdate({ hp: resultHp, tempHp: resultTempHp });
-    } else {
-      const result = calcApplyHealing(prevHp, combatant.maxHp, amount);
-      if (result.hp !== prevHp) {
-        pushHpHistory(combatId, combatant.id, { hp: prevHp, tempHp: prevTempHp, type: 'healing', amount, timestamp: Date.now() });
-        setHistoryVersion(v => v + 1);
-      }
-      onUpdate({ hp: result.hp });
-    }
-  };
-
-  const handleHpAdjustmentChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    setHpAdjustment(value);
-  };
-
-  const applyDamage = () => {
-    const amount = parseInt(hpAdjustment) || 0;
-    if (amount > 0) {
-      adjustHp(-amount);
-      setHpAdjustment('');
-    }
-  };
-
-  const applyHeal = () => {
-    const amount = parseInt(hpAdjustment) || 0;
-    if (amount > 0) {
-      adjustHp(amount);
-      setHpAdjustment('');
-    }
-  };
-
-  const applySetTemp = () => {
-    const amount = parseInt(hpAdjustment) || 0;
-    if (amount > 0) {
-      const currentTempHp = combatant.tempHp ?? 0;
-      const result = calcSetTempHp(currentTempHp, amount);
-      if (result.tempHp === currentTempHp) {
-        setHpAdjustment('');
-        return;
-      }
-      pushHpHistory(combatId, combatant.id, {
-        hp: combatant.hp,
-        tempHp: currentTempHp,
-        type: 'tempHp',
-        amount,
-        timestamp: Date.now(),
-      });
-      setHistoryVersion(v => v + 1);
-      onUpdate({ tempHp: result.tempHp });
-      setHpAdjustment('');
-    }
-  };
-
-  const undoHpChange = () => {
-    const entry = popHpHistory(combatId, combatant.id);
-    if (!entry) return;
-    setHistoryVersion(v => v + 1);
-    onUpdate({ hp: entry.hp, tempHp: entry.tempHp });
-  };
-
-  const addCondition = () => {
-    const name = prompt('Condition name:');
-    if (!name) return;
-
-    const durationStr = prompt('Duration in rounds (leave empty for permanent):');
-    const duration = durationStr ? parseInt(durationStr) : undefined;
-
-    const newCondition: StatusCondition = {
-      id: crypto.randomUUID(),
-      name,
-      description: '',
-      duration,
-    };
-
-    onUpdate({
-      conditions: [...combatant.conditions, newCondition],
-    });
-  };
-
-  const removeCondition = (conditionId: string) => {
-    onUpdate({
-      conditions: combatant.conditions.filter(c => c.id !== conditionId),
-    });
-  };
-
-  const applyDamageToTarget = () => {
-    if (!selectedTargetId || !damageInput || !onUpdateCombatant) return;
-
-    const damage = parseInt(damageInput);
-    if (isNaN(damage) || damage <= 0) {
-      alert('Please enter a valid damage amount');
-      return;
-    }
-
-    const target = allCombatants?.find(c => c.id === selectedTargetId);
-    if (target) {
-      const targetHp = target.hp;
-      const targetTempHp = target.tempHp ?? 0;
-      let resultHp: number;
-      let resultTempHp: number;
-      if (targetDamageType) {
-        const result = calcApplyDamageWithType(
-          targetHp,
-          targetTempHp,
-          damage,
-          targetDamageType,
-          {
-            damageResistances: target.damageResistances,
-            damageImmunities: target.damageImmunities,
-            damageVulnerabilities: target.damageVulnerabilities,
-            activeDamageEffects: target.activeDamageEffects,
-          },
-        );
-        resultHp = result.hp;
-        resultTempHp = result.tempHp;
-      } else {
-        const result = calcApplyDamage(targetHp, targetTempHp, damage);
-        resultHp = result.hp;
-        resultTempHp = result.tempHp;
-      }
-      if (resultHp !== targetHp || resultTempHp !== targetTempHp) {
-        pushHpHistory(combatId, target.id, { hp: targetHp, tempHp: targetTempHp, type: 'damage', amount: damage, timestamp: Date.now() });
-      }
-      onUpdateCombatant(selectedTargetId, { hp: resultHp, tempHp: resultTempHp });
-    }
-
-    setDamageInput('');
-    setTargetDamageType('');
-    setSelectedTargetId(null);
-    setTargetActionMode(null);
-  };
-
-  const addConditionToTarget = () => {
-    if (!selectedTargetId || !newCondition || !onUpdateCombatant) return;
-
-    const target = allCombatants?.find(c => c.id === selectedTargetId);
-    if (target) {
-      const condition: StatusCondition = {
-        id: crypto.randomUUID(),
-        name: newCondition,
-        description: '',
-        duration: conditionDuration ? parseInt(conditionDuration) : undefined,
-      };
-
-      onUpdateCombatant(selectedTargetId, {
-        conditions: [...target.conditions, condition],
-      });
-    }
-
-    setNewCondition('');
-    setConditionDuration('');
-    setSelectedTargetId(null);
-    setTargetActionMode(null);
-  };
-
-  const tempHp = combatant.tempHp ?? 0;
-  const hpTotal = combatant.maxHp + tempHp;
-  const hpPercent = hpTotal > 0 ? (combatant.hp / hpTotal) * 100 : 0;
-  const tempHpPercent = hpTotal > 0 ? (tempHp / hpTotal) * 100 : 0;
-  const hpColor = combatant.maxHp > 0 ? ((combatant.hp / combatant.maxHp) > 0.5 ? 'bg-green-500' : (combatant.hp / combatant.maxHp) > 0.25 ? 'bg-yellow-500' : 'bg-red-500') : 'bg-red-500';
-
-  // Derived damage modifier fields for display and active effects panel
-  const statResistances = combatant.damageResistances ?? [];
-  const statImmunities = combatant.damageImmunities ?? [];
-  const statVulnerabilities = combatant.damageVulnerabilities ?? [];
-  const activeEffects = combatant.activeDamageEffects ?? [];
-  const hasDamageModifiers = statResistances.length > 0 || statImmunities.length > 0 || statVulnerabilities.length > 0 || activeEffects.length > 0;
-
-  // Background gradient based on combatant type - stronger fade from left to right
-  const bgStyle = combatant.type === 'player'
-    ? { backgroundImage: 'linear-gradient(to right, rgba(96, 165, 250, 0.18), rgba(96, 165, 250, 0.02))' }
-    : { backgroundImage: 'linear-gradient(to right, rgba(239, 68, 68, 0.18), rgba(239, 68, 68, 0.02))' };
-
-  return (
-    <div style={bgStyle} className={`rounded-lg px-4 py-4 ${isActive ? 'border-2 border-yellow-500' : 'border border-gray-700'}`}>
-      <div className="flex justify-between items-start">
-        <div className="flex-1">
-          <div className="flex items-center gap-4 mb-2">
-            <div className="flex items-center gap-2">
-              <button
-                onClick={(e) => {
-                  const rect = (e.currentTarget as HTMLButtonElement).getBoundingClientRect();
-                  onShowDetails?.(combatant.id, {
-                    top: rect.bottom,
-                    left: rect.left,
-                  });
-                }}
-                className="hover:opacity-80 transition-opacity"
-                title={`See full ${combatant.type === 'player' ? 'Character' : 'Monster'} information`}
-                type="button"
-                data-testid="combatant-detail-toggle"
-              >
-                <svg
-                  className="w-5 h-5 text-gray-400 hover:text-gray-300 cursor-pointer"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-7-4a1 1 0 11-2 0 1 1 0 012 0zM9 9a1 1 0 000 2v3a1 1 0 001 1h1a1 1 0 100-2v-3a1 1 0 00-1-1H9z"
-                    clipRule="evenodd"
-                  />
-                </svg>
-              </button>
-              <h3 className="text-xl font-semibold">{combatant.name} {combatant.hp <= 0 && '☠️'}</h3>
-              <button
-                onClick={(e) => {
-                  const rect = (e.currentTarget as HTMLButtonElement).getBoundingClientRect();
-                  onShowRemoveConfirm?.(combatant.id, {
-                    top: rect.bottom,
-                    left: rect.left,
-                  });
-                }}
-                className="text-red-500 hover:text-red-400 text-xl leading-none"
-                title="Remove combatant"
-              >
-                ✕
-              </button>
-              {isActive && onNextTurn && (
-                <button
-                  onClick={onNextTurn}
-                  className="px-2 py-1 rounded text-xs bg-yellow-600 hover:bg-yellow-700 animate-pulse font-semibold"
-                >
-                  Current Turn (done)
-                </button>
-              )}
-            </div>
-            {!isActive && <div className="w-40"></div>}
-            <div className="flex items-center gap-2">
-              <p className="text-xs text-gray-400">AC</p>
-              <p className="text-lg font-bold">{combatant.ac}</p>
-            </div>
-            <span className="text-sm text-gray-400 whitespace-nowrap">Hit Points:</span>
-            <span className="text-lg font-bold">
-              Current: <span className={hpColor === 'bg-green-500' ? 'text-green-500' : hpColor === 'bg-yellow-500' ? 'text-yellow-500' : 'text-red-500'}>{combatant.hp}</span> Max: {combatant.maxHp}{tempHp > 0 && <span className="text-blue-400"> +{tempHp} tmp</span>}
-            </span>
-            {(combatant.legendaryActionCount ?? 0) > 0 && (
-              <span
-                className="text-sm font-semibold text-amber-400 whitespace-nowrap"
-                data-testid="legendary-action-badge"
-              >
-                ⚡ {combatant.legendaryActionsRemaining ?? combatant.legendaryActionCount}/{combatant.legendaryActionCount}
-              </span>
-            )}
-            <input
-              type="number"
-              placeholder="0"
-              value={hpAdjustment}
-              onChange={handleHpAdjustmentChange}
-              onKeyPress={(e) => {
-                if (e.key === 'Enter') {
-                  if (e.shiftKey) {
-                    isTempMode ? applySetTemp() : applyHeal();
-                  } else {
-                    applyDamage();
-                  }
-                }
-              }}
-              className="w-14 bg-gray-700 rounded px-2 py-1 text-xs text-center text-white"
-            />
-            <select
-              value={selectedDamageType}
-              onChange={(e) => setSelectedDamageType(e.target.value as DamageType | '')}
-              className="bg-gray-700 rounded px-1 py-1 text-xs text-white border border-gray-600"
-              title="Damage type (for resistance/immunity/vulnerability)"
-              aria-label="Damage type (for resistance/immunity/vulnerability)"
-            >
-              <option value="">Type</option>
-              {Object.entries(DAMAGE_TYPE_GROUPS).map(([group, types]) => (
-                <optgroup key={group} label={group}>
-                  {types.map(t => (
-                    <option key={t} value={t}>{t.charAt(0).toUpperCase() + t.slice(1)}</option>
-                  ))}
-                </optgroup>
-              ))}
-            </select>
-            <button
-              onClick={applyDamage}
-              title={selectedDamageType ? `Apply ${selectedDamageType} damage (with resistances)` : "Apply damage (Enter)"}
-              className={`px-2 py-1 rounded text-xs ${selectedDamageType ? 'bg-orange-600 hover:bg-orange-700' : 'bg-red-600 hover:bg-red-700'}`}
-            >
-              Damage
-            </button>
-            <button
-              onClick={isTempMode ? applySetTemp : applyHeal}
-              title={isTempMode ? "Set temporary HP" : "Apply healing (Shift+Enter)"}
-              className="bg-green-600 hover:bg-green-700 px-2 py-1 rounded text-xs"
-            >
-              {isTempMode ? 'Set Temp' : 'Heal'}
-            </button>
-            <label className="flex items-center gap-1 text-xs text-gray-300 cursor-pointer">
-              <input
-                type="checkbox"
-                checked={isTempMode}
-                onChange={(e) => setIsTempMode(e.target.checked)}
-                className="cursor-pointer"
-              />
-              Temp
-            </label>
-            <button
-              onClick={undoHpChange}
-              disabled={getHpHistoryStack(combatId, combatant.id).length === 0}
-              title="Undo last HP change"
-              data-testid="undo-hp-change"
-              className="px-2 py-1 rounded text-xs bg-gray-600 hover:bg-gray-500 disabled:opacity-40 disabled:cursor-not-allowed"
-            >
-              Undo HP
-            </button>
-            <div className="flex items-center gap-2 ml-auto pr-4">
-              <button
-                onClick={() => onSetInitiative?.(combatant.id)}
-                className="flex items-center gap-1 hover:opacity-80 cursor-pointer transition-opacity"
-              >
-                <p className="text-xs text-gray-400">Initiative</p>
-                <p className="text-lg font-bold">{combatant.initiative}</p>
-                {combatant.initiativeRoll && (
-                  <p className="text-xs text-gray-500 whitespace-nowrap">
-                    {combatant.initiativeRoll.method === 'rolled'
-                      ? [
-                          combatant.initiativeRoll.advantage
-                            ? `d20:${combatant.initiativeRoll.roll}↑${combatant.initiativeRoll.altRoll != null ? ` (dropped:${combatant.initiativeRoll.altRoll})` : ''}`
-                            : `d20:${combatant.initiativeRoll.roll}`,
-                          `+${combatant.initiativeRoll.bonus}`,
-                          combatant.initiativeRoll.flatBonus
-                            ? `${combatant.initiativeRoll.flatBonus > 0 ? '+' : ''}${combatant.initiativeRoll.flatBonus}`
-                            : null,
-                        ].filter(Boolean).join('')
-                      : [
-                          combatant.initiativeRoll.roll != null
-                            ? String(combatant.initiativeRoll.roll)
-                            : null,
-                          combatant.initiativeRoll.bonus !== 0
-                            ? `${combatant.initiativeRoll.bonus > 0 ? '+' : ''}${combatant.initiativeRoll.bonus}`
-                            : null,
-                          combatant.initiativeRoll.flatBonus
-                            ? `${combatant.initiativeRoll.flatBonus > 0 ? '+' : ''}${combatant.initiativeRoll.flatBonus}`
-                            : null,
-                        ].filter((part): part is string => part != null).join('') || 'Manual'}
-                  </p>
-                )}
-              </button>
-            </div>
-          </div>
-
-          <div className="w-4/5 bg-gray-700 rounded-full h-2 flex overflow-hidden">
-            <div className={`${hpColor} h-2 transition-all`} data-testid="health-bar" style={{ width: `${hpPercent}%` }} />
-            {tempHp > 0 && (
-              <div className="bg-blue-400 h-2 transition-all" data-testid="temp-hp-bar" style={{ width: `${tempHpPercent}%` }} />
-            )}
-          </div>
-
-          {/* Active Damage Effects */}
-          <div className="mt-1 mb-1">
-                <div className="flex items-center gap-2 flex-wrap">
-                  {statImmunities.map(t => (
-                    <span key={`imm-${t}`} className="text-xs bg-purple-900 text-purple-200 px-1.5 py-0.5 rounded font-semibold" title="Immunity (from stats)">
-                      IMM: {t}
-                    </span>
-                  ))}
-                  {statResistances.map(t => (
-                    <span key={`res-${t}`} className="text-xs bg-green-900 text-green-200 px-1.5 py-0.5 rounded" title="Resistance (from stats)">
-                      RES: {t}
-                    </span>
-                  ))}
-                  {statVulnerabilities.map(t => (
-                    <span key={`vuln-${t}`} className="text-xs bg-red-900 text-red-200 px-1.5 py-0.5 rounded" title="Vulnerability (from stats)">
-                      VULN: {t}
-                    </span>
-                  ))}
-                  {activeEffects.map(e => (
-                    <span key={`active-${e.type}-${e.kind}`} className="text-xs bg-yellow-800 text-yellow-200 px-1.5 py-0.5 rounded flex items-center gap-1" title={e.label}>
-                      {e.kind === 'immunity' ? 'IMM' : e.kind === 'resistance' ? 'RES' : 'VULN'}: {e.type} ✱
-                      <button
-                        onClick={() => onUpdate({ activeDamageEffects: removeActiveDamageEffects(activeEffects, e.type, e.kind) })}
-                        className="ml-0.5 text-yellow-300 hover:text-white leading-none"
-                        title={`Remove ${e.label}`}
-                        aria-label={`Remove ${e.label}`}
-                      >✕</button>
-                    </span>
-                  ))}
-                  <button
-                    onClick={() => { setShowEffectsPanel(v => { if (v) setPendingPreset(null); return !v; }); }}
-                    className="text-xs text-teal-400 hover:text-teal-300 underline"
-                    title="Add temporary combat damage effects"
-                  >
-                    {showEffectsPanel ? 'Hide effects' : '+ Add effect'}
-                  </button>
-                </div>
-                {showEffectsPanel && (
-                  <div className="mt-2 bg-gray-800 border border-teal-700 rounded p-3 space-y-2">
-                    {pendingPreset ? (
-                      <>
-                        <p className="text-xs text-teal-300 font-semibold">{pendingPreset.label}: choose a damage type</p>
-                        <div className="flex flex-wrap gap-1">
-                          {pendingPreset.choices.map(t => (
-                            <button
-                              key={t}
-                              onClick={() => {
-                                const effect: ActiveDamageEffect = { type: t, kind: pendingPreset.kind, label: pendingPreset.label };
-                                onUpdate({ activeDamageEffects: mergeActiveDamageEffects(activeEffects, [effect]) });
-                                setPendingPreset(null);
-                                setShowEffectsPanel(false);
-                              }}
-                              className="text-xs bg-teal-800 hover:bg-teal-700 text-teal-200 border border-teal-600 px-2 py-0.5 rounded capitalize"
-                            >
-                              {t}
-                            </button>
-                          ))}
-                          <button
-                            onClick={() => setPendingPreset(null)}
-                            className="text-xs text-gray-400 hover:text-gray-300 px-1"
-                          >
-                            ← Back
-                          </button>
-                        </div>
-                      </>
-                    ) : (
-                      <>
-                        <p className="text-xs text-teal-300 font-semibold">Apply a combat damage effect:</p>
-                        <div className="flex flex-wrap gap-2">
-                          {DAMAGE_EFFECT_PRESETS.map(preset => {
-                            const hasNullType = preset.effects.some(e => e.type === null);
-                            return (
-                              <button
-                                key={preset.id}
-                                onClick={() => {
-                                  if (hasNullType) {
-                                    // Preset requires user to pick a type — open secondary picker
-                                    const nullEffect = preset.effects.find(e => e.type === null)!;
-                                    const choices: DamageType[] = nullEffect.choicesLimited ??
-                                      ([...DAMAGE_TYPE_GROUPS.Physical, ...DAMAGE_TYPE_GROUPS.Elemental, ...DAMAGE_TYPE_GROUPS['Energy & Planar'], ...DAMAGE_TYPE_GROUPS.Other]);
-                                    setPendingPreset({ label: preset.label, kind: nullEffect.kind, choices });
-                                  } else {
-                                    const newEffects = preset.effects.map(e => ({
-                                      type: e.type as DamageType,
-                                      kind: e.kind,
-                                      label: preset.label,
-                                    }));
-                                    onUpdate({ activeDamageEffects: mergeActiveDamageEffects(activeEffects, newEffects) });
-                                    setShowEffectsPanel(false);
-                                  }
-                                }}
-                                className="text-xs bg-teal-700 hover:bg-teal-600 text-white px-2 py-1 rounded"
-                                title={preset.description}
-                              >
-                                {preset.label}{hasNullType ? ' →' : ''}
-                              </button>
-                            );
-                          })}
-                        </div>
-                      </>
-                    )}
-                    <div className="flex items-center gap-2 pt-1">
-                      <span className="text-xs text-gray-400">Custom:</span>
-                      {(['resistance', 'immunity', 'vulnerability'] as const).map(kind => (
-                        selectedDamageType ? (
-                          <button
-                            key={kind}
-                            onClick={() => {
-                              const effect: ActiveDamageEffect = { type: selectedDamageType, kind, label: `${kind} (${selectedDamageType})` };
-                              onUpdate({ activeDamageEffects: mergeActiveDamageEffects(activeEffects, [effect]) });
-                              setShowEffectsPanel(false);
-                            }}
-                            className="text-xs bg-gray-700 hover:bg-gray-600 border border-teal-600 text-teal-200 px-2 py-1 rounded capitalize"
-                          >
-                            {kind} ({selectedDamageType})
-                          </button>
-                        ) : null
-                      ))}
-                      {!selectedDamageType && <span className="text-xs text-gray-500 italic">Select a damage type above first</span>}
-                    </div>
-                  </div>
-                )}
-              </div>
-
-          {combatant.conditions.length > 0 && (
-            <div className="mb-2">
-              <button
-                onClick={() => setShowConditions(!showConditions)}
-                className="text-sm text-blue-400 hover:text-blue-300"
-              >
-                Conditions ({combatant.conditions.length})
-              </button>
-              {showConditions && (
-                <div className="mt-2 space-y-1">
-                  {combatant.conditions.map(condition => (
-                    <div key={condition.id} className="bg-gray-700 rounded px-2 py-1 text-sm flex justify-between items-center">
-                      <span>
-                        {condition.name}
-                        {condition.duration && ` (${condition.duration} rounds)`}
-                      </span>
-                      <button
-                        onClick={() => removeCondition(condition.id)}
-                        className="text-red-400 hover:text-red-300 text-xs"
-                      >
-                        Remove
-                      </button>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
-          )}
-
-          {combatant.targetIds && combatant.targetIds.length > 0 && (
-            <div className="mb-2">
-              <div className="flex flex-wrap items-center gap-2">
-                <span className="text-sm text-purple-400 font-semibold">Targets:</span>
-                {combatant.targetIds.map(targetId => {
-                  const target = allCombatants?.find(c => c.id === targetId);
-                  return target ? (
-                    <div key={targetId} className="relative inline-block">
-                      <button
-                        onClick={() => {
-                          setSelectedTargetId(targetId);
-                          setTargetActionMode(null);
-                        }}
-                        onMouseEnter={() => setHoveredTargetId(targetId)}
-                        onMouseLeave={() => setHoveredTargetId(null)}
-                        className={`px-2 py-1 rounded text-xs font-semibold cursor-pointer transition-all hover:opacity-80 ${target.type === 'player' ? 'bg-blue-600 hover:bg-blue-700 text-blue-100' : 'bg-red-600 hover:bg-red-700 text-red-100'}`}
-                      >
-                        {target.name}
-                      </button>
-                      {hoveredTargetId === targetId && (
-                        <div className="absolute bottom-full left-0 mb-2 bg-gray-900 border border-gray-700 rounded shadow-lg pointer-events-none z-50 min-w-max">
-                          <div className="px-3 py-2 space-y-1">
-                            <div className="text-xs text-gray-400">
-                              <div>AC: {target.ac}</div>
-                              <div className="flex items-center gap-1">
-                                HP: <span className={target.hp === 0 ? 'text-red-400' : 'text-gray-300'}>{target.hp}/{target.maxHp}</span>
-                                {target.hp === 0 && <span className="text-red-400 text-lg">☠</span>}
-                              </div>
-                            </div>
-                            {target.conditions.length > 0 && (
-                              <div className="text-xs space-y-1 pt-1">
-                                {target.conditions.map((condition) => (
-                                  <div key={condition.id} className="text-yellow-400">
-                                    • {condition.name}
-                                    {condition.duration && ` (${condition.duration})`}
-                                  </div>
-                                ))}
-                              </div>
-                            )}
-                            <div className="border-t border-gray-600 mt-2 pt-2 text-xs text-gray-400 italic">
-                              Click to apply damage or add condition
-                            </div>
-                          </div>
-                        </div>
-                      )}
-                    </div>
-                  ) : null;
-                })}
-              </div>
-            </div>
-          )}
-
-          {combatant.notes && (
-            <p className="text-sm text-gray-400 italic">{combatant.notes}</p>
-          )}
-        </div>
-
-        <div className="flex flex-col gap-2">
-          <button
-            onClick={() => setShowTargeting(!showTargeting)}
-            className="bg-orange-600 hover:bg-orange-700 px-2 py-1 rounded text-xs"
-            title="Set targets for this combatant"
-          >
-            Add Target(s)
-          </button>
-          <button
-            onClick={addCondition}
-            className="bg-purple-600 hover:bg-purple-700 px-2 py-1 rounded text-xs"
-          >
-            Add Condition
-          </button>
-        </div>
-      </div>
-
-      {showTargeting && allCombatants && (
-        <div className="mt-4 bg-gray-800 rounded p-4 border border-purple-600">
-          <div className="flex justify-between items-center mb-3">
-            <h4 className="text-sm font-semibold text-purple-300">Select targets for {combatant.name}</h4>
-            <button
-              onClick={() => setShowTargeting(false)}
-              className="text-gray-400 hover:text-gray-300 text-lg"
-            >
-              ✕
-            </button>
-          </div>
-          <div className="grid grid-cols-2 gap-4">
-            {/* Monsters Column */}
-            <div>
-              <h5 className="text-xs font-semibold text-red-300 mb-2 uppercase">Enemies</h5>
-              <div className="space-y-2">
-                {allCombatants
-                  .filter(c => c.id !== combatant.id && c.type !== 'player' && c.type !== 'lair')
-                  .sort((a, b) => a.name.localeCompare(b.name))
-                  .map(target => (
-                    <label key={target.id} className="flex items-center gap-2 cursor-pointer">
-                      <input
-                        type="checkbox"
-                        checked={combatant.targetIds?.includes(target.id) ?? false}
-                        onChange={(e) => {
-                          if (e.target.checked) {
-                            onUpdate({
-                              targetIds: [...(combatant.targetIds ?? []), target.id],
-                            });
-                          } else {
-                            onUpdate({
-                              targetIds: (combatant.targetIds ?? []).filter(id => id !== target.id),
-                            });
-                          }
-                        }}
-                        className="w-4 h-4 rounded bg-gray-700 border border-gray-600 cursor-pointer"
-                      />
-                      <span className="text-sm text-red-300">
-                        {target.name}
-                      </span>
-                    </label>
-                  ))}
-              </div>
-            </div>
-            
-            {/* Party Column */}
-            <div>
-              <h5 className="text-xs font-semibold text-blue-300 mb-2 uppercase">Party</h5>
-              <div className="space-y-2">
-                {allCombatants
-                  .filter(c => c.id !== combatant.id && c.type === 'player')
-                  .sort((a, b) => a.name.localeCompare(b.name))
-                  .map(target => (
-                    <label key={target.id} className="flex items-center gap-2 cursor-pointer">
-                      <input
-                        type="checkbox"
-                        checked={combatant.targetIds?.includes(target.id) ?? false}
-                        onChange={(e) => {
-                          if (e.target.checked) {
-                            onUpdate({
-                              targetIds: [...(combatant.targetIds ?? []), target.id],
-                            });
-                          } else {
-                            onUpdate({
-                              targetIds: (combatant.targetIds ?? []).filter(id => id !== target.id),
-                            });
-                          }
-                        }}
-                        className="w-4 h-4 rounded bg-gray-700 border border-gray-600 cursor-pointer"
-                      />
-                      <span className="text-sm text-blue-300">
-                        {target.name}
-                      </span>
-                    </label>
-                  ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Target Action Modal */}
-      {selectedTargetId && allCombatants && (
-        <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50">
-          <div className="bg-gray-800 border border-gray-700 rounded-lg p-6 max-w-sm mx-auto">
-            {(() => {
-              const target = allCombatants.find(c => c.id === selectedTargetId);
-              if (!target) return null;
-
-              return (
-                <>
-                  <h3 className="text-lg font-semibold mb-4 text-white">
-                    {target.name}
-                  </h3>
-
-                  {!targetActionMode ? (
-                    <div className="space-y-3">
-                      <p className="text-sm text-gray-400 mb-4">
-                        HP: {target.hp}/{target.maxHp} | AC: {target.ac}
-                      </p>
-                      <button
-                        onClick={() => setTargetActionMode('damage')}
-                        className="w-full bg-red-600 hover:bg-red-700 px-4 py-2 rounded text-white font-semibold transition-colors"
-                      >
-                        Apply Damage
-                      </button>
-                      <button
-                        onClick={() => setTargetActionMode('condition')}
-                        className="w-full bg-purple-600 hover:bg-purple-700 px-4 py-2 rounded text-white font-semibold transition-colors"
-                      >
-                        Add Condition
-                      </button>
-                      <button
-                        onClick={() => {
-                          setSelectedTargetId(null);
-                          setTargetActionMode(null);
-                        }}
-                        className="w-full bg-gray-700 hover:bg-gray-600 px-4 py-2 rounded text-white font-semibold transition-colors"
-                      >
-                        Cancel
-                      </button>
-                    </div>
-                  ) : targetActionMode === 'damage' ? (
-                    <div className="space-y-3">
-                      <input
-                        type="number"
-                        min="0"
-                        value={damageInput}
-                        onChange={(e) => setDamageInput(e.target.value)}
-                        placeholder="Damage amount"
-                        className="w-full bg-gray-700 rounded px-3 py-2 text-white focus:outline-none focus:ring-2 focus:ring-red-500"
-                        autoFocus
-                      />
-                      <select
-                        value={targetDamageType}
-                        onChange={(e) => setTargetDamageType(e.target.value as DamageType | '')}
-                        className="w-full bg-gray-700 rounded px-3 py-2 text-white border border-gray-600"
-                        title="Damage type (applies target's resistances, immunities, and vulnerabilities)"
-                        aria-label="Damage type (applies target's resistances, immunities, and vulnerabilities)"
-                      >
-                        <option value="">No damage type</option>
-                        {Object.entries(DAMAGE_TYPE_GROUPS).map(([group, types]) => (
-                          <optgroup key={group} label={group}>
-                            {types.map(t => (
-                              <option key={t} value={t}>{t.charAt(0).toUpperCase() + t.slice(1)}</option>
-                            ))}
-                          </optgroup>
-                        ))}
-                      </select>
-                      <div className="flex gap-2">
-                        <button
-                          onClick={applyDamageToTarget}
-                          className={`flex-1 px-3 py-2 rounded text-white font-semibold transition-colors ${targetDamageType ? 'bg-orange-600 hover:bg-orange-700' : 'bg-red-600 hover:bg-red-700'}`}
-                        >
-                          Apply{targetDamageType ? ` (${targetDamageType})` : ''}
-                        </button>
-                        <button
-                          onClick={() => setTargetActionMode(null)}
-                          className="flex-1 bg-gray-700 hover:bg-gray-600 px-3 py-2 rounded text-white font-semibold transition-colors"
-                        >
-                          Back
-                        </button>
-                      </div>
-                    </div>
-                  ) : (
-                    <div className="space-y-3">
-                      <input
-                        type="text"
-                        value={newCondition}
-                        onChange={(e) => setNewCondition(e.target.value)}
-                        placeholder="Condition name"
-                        className="w-full bg-gray-700 rounded px-3 py-2 text-white focus:outline-none focus:ring-2 focus:ring-purple-500"
-                        autoFocus
-                      />
-                      <input
-                        type="number"
-                        min="0"
-                        value={conditionDuration}
-                        onChange={(e) => setConditionDuration(e.target.value)}
-                        placeholder="Duration in rounds (optional)"
-                        className="w-full bg-gray-700 rounded px-3 py-2 text-white focus:outline-none focus:ring-2 focus:ring-purple-500"
-                      />
-                      <div className="flex gap-2">
-                        <button
-                          onClick={addConditionToTarget}
-                          className="flex-1 bg-purple-600 hover:bg-purple-700 px-3 py-2 rounded text-white font-semibold transition-colors"
-                        >
-                          Add
-                        </button>
-                        <button
-                          onClick={() => setTargetActionMode(null)}
-                          className="flex-1 bg-gray-700 hover:bg-gray-600 px-3 py-2 rounded text-white font-semibold transition-colors"
-                        >
-                          Back
-                        </button>
-                      </div>
-                    </div>
-                  )}
-                </>
-              );
-            })()}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
 export default function CombatPage() {
-  return (
-    <ProtectedRoute>
-      <CombatContent />
-    </ProtectedRoute>
-  );
+  return <ProtectedRoute><CombatContent /></ProtectedRoute>;
 }
-
-interface InitiativeEntryProps {
-  combatant: CombatantState;
-  onSet: (initiativeRoll: InitiativeRoll) => void;
-  onClose?: () => void; // optional: close the edit form (only valid when initiative exists)
-  onSettingsChange?: (advantage: boolean, flatBonus: number) => void;
-}
-
-export function InitiativeEntry({ combatant, onSet, onClose, onSettingsChange }: InitiativeEntryProps) {
-  const [entryMode, setEntryMode] = useState<'roll' | 'dice' | 'total'>('roll');
-  const [diceRoll, setDiceRoll] = useState('');
-  const [totalValue, setTotalValue] = useState('');
-  const [advantage, setAdvantage] = useState(combatant.initiativeAdvantage ?? false);
-  const [flatBonus, setFlatBonus] = useState(combatant.initiativeFlatBonus ?? 0);
-
-  // Close with Escape only when there is an existing initiative and an onClose handler
-  useEffect(() => {
-    if (!combatant.initiativeRoll || !onClose) return;
-    const handler = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') onClose();
-    };
-    document.addEventListener('keydown', handler);
-    return () => document.removeEventListener('keydown', handler);
-  }, [combatant.initiativeRoll, onClose]);
-
-  const handleRoll = () => {
-    onSet(buildInitiativeRoll({ ...combatant, initiativeAdvantage: advantage, initiativeFlatBonus: flatBonus }));
-  };
-
-  const handleDiceEntry = () => {
-    const roll = parseInt(diceRoll) || 0;
-    if (roll < 1 || roll > 20) {
-      alert('Dice roll must be between 1 and 20');
-      return;
-    }
-
-    const bonus = getDexInitiativeBonus(combatant);
-    const effectiveFlatBonus = flatBonus !== 0 ? flatBonus : undefined;
-    const total = roll + bonus + (flatBonus ?? 0);
-
-    onSet({
-      roll,
-      bonus,
-      total,
-      method: 'manual',
-      ...(effectiveFlatBonus !== undefined && { flatBonus: effectiveFlatBonus }),
-    });
-  };
-
-  const handleTotalEntry = () => {
-    const total = parseInt(totalValue) || 0;
-    if (total < 0) {
-      alert('Initiative must be 0 or greater');
-      return;
-    }
-
-    const effectiveFlatBonus = flatBonus !== 0 ? flatBonus : undefined;
-
-    onSet({
-      roll: total,
-      bonus: 0,
-      total: total + (flatBonus ?? 0),
-      method: 'manual',
-      ...(effectiveFlatBonus !== undefined && { flatBonus: effectiveFlatBonus }),
-    });
-  };
-
-  return (
-    <div className="relative bg-gray-800 rounded-lg p-4 border border-gray-700">
-      {/* Close button only for entries that already have an initiative and when parent provided onClose */}
-      {combatant.initiativeRoll && onClose && (
-        <button
-          onClick={onClose}
-          aria-label={`Close initiative editor for ${combatant.name}`}
-          className="absolute -top-3 -right-3 w-7 h-7 rounded-full bg-gray-800 border border-gray-700 text-gray-400 hover:text-gray-200 flex items-center justify-center text-lg"
-          type="button"
-        >
-          ×
-        </button>
-      )}
-
-      <div className="flex justify-between items-start mb-4">
-        <div className="flex flex-col md:flex-row items-start gap-4">
-              <div>
-            <div className="flex items-center gap-2">
-              <span className="text-sm text-gray-400">Set Initiative:</span>
-              <h3 className="text-lg font-semibold">{combatant.name}</h3>
-            </div>
-            <div className="flex justify-end mt-2">
-              <span
-                className={`px-3 py-1 rounded-full text-xs font-semibold ${
-                combatant.type === "player"
-                  ? "bg-blue-600 text-blue-100"
-                  : "bg-red-600 text-red-100"
-                }`}
-              >
-                {combatant.type === "player" ? "Character" : "Monster"}
-              </span>
-            </div>
-            </div>
-
-          {/* Buttons: stacked on mobile (full width), horizontal on md+ */}
-          <div className="flex flex-col md:flex-row gap-2 md:pl-3 w-full md:w-auto">
-            <button
-              onClick={() => {
-                setEntryMode("roll");
-                handleRoll();
-              }}
-              className={`w-full md:w-28 px-2 py-2 rounded text-sm flex-none ${
-                entryMode === "roll"
-                  ? "bg-blue-600"
-                  : "bg-gray-700 hover:bg-gray-600"
-              }`}
-            >
-              Roll d20
-            </button>
-            <button
-              onClick={() => setEntryMode("dice")}
-              className={`w-full md:w-40 px-2 py-2 rounded text-sm flex-none ${
-                entryMode === "dice"
-                  ? "bg-purple-600"
-                  : "bg-gray-700 hover:bg-gray-600"
-              }`}
-            >
-              Enter Dice Roll
-            </button>
-            <button
-              onClick={() => setEntryMode("total")}
-              className={`w-full md:w-28 px-2 py-2 rounded text-sm flex-none ${
-                entryMode === "total"
-                  ? "bg-green-600"
-                  : "bg-gray-700 hover:bg-gray-600"
-              }`}
-            >
-              Enter Total
-            </button>
-          </div>
-        </div>
-
-        <div className="flex flex-wrap items-center gap-4 mt-2">
-          <label className="flex items-center gap-2 text-sm text-gray-300 cursor-pointer">
-            <input
-              type="checkbox"
-              checked={advantage}
-              onChange={(e) => {
-                const next = e.target.checked;
-                setAdvantage(next);
-                onSettingsChange?.(next, flatBonus);
-              }}
-              className="w-4 h-4 accent-blue-500"
-            />
-            Advantage
-          </label>
-          <div className="flex items-center gap-1">
-            <span className="text-sm text-gray-400">Flat bonus:</span>
-            <input
-              type="number"
-              step={1}
-              value={flatBonus}
-              onChange={(e) => {
-                const next = e.target.value === '' || !Number.isFinite(e.target.valueAsNumber)
-                  ? 0
-                  : Math.trunc(e.target.valueAsNumber);
-                setFlatBonus(next);
-                onSettingsChange?.(advantage, next);
-              }}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  e.currentTarget.blur();
-                }
-              }}
-              className="w-16 bg-gray-700 rounded px-2 py-1 text-sm text-white"
-              aria-label="Flat initiative bonus"
-            />
-            {flatBonus !== 0 && (
-              <button
-                type="button"
-                onClick={() => {
-                  setFlatBonus(0);
-                  onSettingsChange?.(advantage, 0);
-                }}
-                className="text-gray-400 hover:text-gray-200 text-sm px-1"
-                aria-label="Clear flat bonus"
-              >
-                ✕
-              </button>
-            )}
-          </div>
-        </div>
-
-        <div className="flex items-start gap-2">
-          {entryMode === "dice" && (
-            <div className="flex gap-2">
-              <input
-                type="number"
-                min="1"
-                max="20"
-                value={diceRoll}
-                onChange={(e) => setDiceRoll(e.target.value)}
-                placeholder="1-20"
-                className="flex-1 bg-gray-700 rounded px-3 py-2 text-white"
-              />
-              <button
-                onClick={handleDiceEntry}
-                className="bg-purple-600 hover:bg-purple-700 w-20 px-2 py-2 rounded"
-              >
-                Set
-              </button>
-            </div>
-          )}
-
-          {entryMode === "total" && (
-            <div className="flex gap-2">
-              <input
-                type="number"
-                min="0"
-                value={totalValue}
-                onChange={(e) => setTotalValue(e.target.value)}
-                placeholder={flatBonus !== 0 ? `Value (${flatBonus > 0 ? '+' : ''}${flatBonus} bonus applied)` : "Total initiative"}
-                className="flex-1 bg-gray-700 rounded px-3 py-2 text-white"
-              />
-              <button
-                onClick={handleTotalEntry}
-                className="bg-green-600 hover:bg-green-700 w-20 px-2 py-2 rounded"
-              >
-                Set
-              </button>
-            </div>
-          )}
-
-          {combatant.initiativeRoll && (
-            <div className="bg-gray-700 rounded px-3 py-2 text-sm">
-              <p className="text-gray-400">
-                Initiative:{" "}
-                <span className="text-white font-bold">
-                  {combatant.initiativeRoll.total}
-                </span>
-              </p>
-              {combatant.initiativeRoll.method === "rolled" && (
-                <p className="text-gray-500 text-xs">
-                  {combatant.initiativeRoll.advantage ? (
-                    <>d20: {combatant.initiativeRoll.roll}↑{combatant.initiativeRoll.altRoll != null ? ` (dropped: ${combatant.initiativeRoll.altRoll})` : ''}</>
-                  ) : (
-                    <>d20: {combatant.initiativeRoll.roll}</>
-                  )}{" "}
-                  + {combatant.initiativeRoll.bonus}
-                  {combatant.initiativeRoll.flatBonus ? ` ${combatant.initiativeRoll.flatBonus > 0 ? '+' : ''}${combatant.initiativeRoll.flatBonus}` : ""}{" "}
-                  = {combatant.initiativeRoll.total}
-                </p>
-              )}
-              {combatant.initiativeRoll.method === "manual" && (combatant.initiativeRoll.bonus !== 0 || combatant.initiativeRoll.flatBonus) ? (
-                <p className="text-gray-500 text-xs">
-                  {combatant.initiativeRoll.roll}
-                  {combatant.initiativeRoll.bonus !== 0 && ` + ${combatant.initiativeRoll.bonus}`}
-                  {combatant.initiativeRoll.flatBonus ? ` ${combatant.initiativeRoll.flatBonus > 0 ? '+' : ''}${combatant.initiativeRoll.flatBonus}` : ""}
-                  {" "}= {combatant.initiativeRoll.total}
-                </p>
-              ) : null}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
